Update to latest bounded

+++ source/tm/ps_usage_stats/serialize.cpp
@@ -33,7 +33,7 @@ template<typename Key>
 constexpr auto get_used(auto const get, auto predicate) {
 	using Mapped = decltype(get(std::declval<Key>()));
 	auto transformer = [&](Key const name) { return NameValue<Key, Mapped>{name, get(name)}; };
-	return containers::static_vector<NameValue<Key, Mapped>, static_cast<std::size_t>(number_of<Key>)>(
+	return containers::static_vector<NameValue<Key, Mapped>, number_of<Key>>(
 		containers::filter(
 			containers::transform(containers::enum_range<Key>(), transformer),
 			predicate
