+++ source/tm/team_predictor/ev_optimizer/defensive.hpp
@@ -129,7 +129,7 @@ struct DefensiveEVs {
 	}
 private:
 	static constexpr auto maximum_natures = containers::size(containers::enum_range<Nature>());
-	containers::static_vector<DataPoint, static_cast<std::size_t>(maximum_natures * bounded::pow(max_possible_optimized_ivs, 3_bi))> m_container;
+	containers::static_vector<DataPoint, maximum_natures * bounded::pow(max_possible_optimized_ivs, 3_bi)> m_container;
 };
 
 }	// namespace technicalmachine
